+++ tests/test_utils.py
@@ -189,7 +189,7 @@ def test_reparse_args_burst2safe():
     with pytest.raises(ValueError, match='Cannot provide*'):
         utils.reparse_args(args1, 'burst2safe')
 
-    args2 = MockArgs(orbit=123, pols=['VV', 'HH'])
+    args2 = MockArgs(orbit=123)
     with pytest.raises(ValueError, match='Must provide*'):
         utils.reparse_args(args2, 'burst2safe')
 
@@ -205,7 +205,7 @@ def test_reparse_args_burst2stack():
         def __init__(self, **kwargs):
             self.__dict__.update(kwargs)
 
-    args1 = MockArgs(rel_orb=123, pols=['VV', 'HH'])
+    args1 = MockArgs(rel_orb=123)
     with pytest.raises(ValueError, match='Must provide*'):
         utils.reparse_args(args1, 'burst2stack')